Update vote.sol
The 'view' modifier has been added to the 'readCandidate' and 'getCandidatesLength' functions to indicate that they are read-only and don't modify the contract state. This helps optimize gas consumption when calling these functions.

The 'price' and 'votes' variables are initialized directly within the Candidate struct during candidate creation, eliminating the need for separate assignment statements.